Split test_J2_bspline from test_bspline_jastrow.
